Fix bugs in the board's checker, misc other things.
Mainly rewrote large parts of the `check_moves` function in the
`BoardState` class as it was previously unorganized and didn't
correctly check for broken up lines of letters in a single move,
among other things. Modified a few minor parts of the `BoardState` class
including consistent style for class fields and struct names.

Fix a minor bug in the command line interface front end, which didn't
update the board state's "current move vs previous move" move state
when a successful move is made.

Fix a link to Aspell project page in readme.

Add the name of the front end's executable binary to gitignore.